+++ Change Log.md
@@ -11,6 +11,7 @@ Changes by GTFreeFlyer, date TBD
 * Fixed AutoDraw line showing up even when user set it to false
 * Added aircraft type to the label next to the best race line
 * Duplicate best time by racer shown in the info header - FIXED
+* Added new F10 menu category, Top 10 Times. This is in addition to the existing Top 10 Racers.
 * Added more general-purpose flags for mission design options:
    * RacerCrashed, RacerEjected, RacerDied, RacerDisconnected, and NewBestTime
 -------------------------------------------------------------------------------------